+++ src/green/GreenSym.jl
@@ -336,7 +336,7 @@ Interpolation method is by default depending on the grid, but could also be chos
 - 'color2': Target color2
 - 'spaceMethod': Method of interpolation for space. 
 """
-function instant(green::GreenSym2DLR{DT,TT,TGT,SGT}, space, color1::Int, color2::Int=color1; spaceMethod::SM = DEFAULTINTERP) where {DT,TT,TGT,SGT,SM}
+function instant(; green::Union{Green2DLR{DT,TT,TGT,SGT},GreenSym2DLR{DT,TT,TGT,SGT}}, space, color1::Int, color2::Int, spaceMethod::SM) where {DT,TT,TGT,SGT,SM}
     if isempty(green.instant)
         error("Instant Green's function can not be empty!")
     else
@@ -347,11 +347,17 @@ function instant(green::GreenSym2DLR{DT,TT,TGT,SGT}, space, color1::Int, color2:
     end
 end
 
-function instant(green::GreenSym2DLR{DT,TT,TGT,SGT}, space; spaceMethod::SM = DEFAULTINTERP) where {DT,TT,TGT,SGT,SM}
-    return instant(green, space, 1, 1; spaceMethod)
+function instant(green::Union{Green2DLR{DT,TT,TGT,SGT},GreenSym2DLR{DT,TT,TGT,SGT}}, space, color1::Int=1, color2::Int=color1, spaceMethod::SM = DEFAULTINTERP) where {DT,TT,TGT,SGT,SM}
+    return instant(; green=green, space = space, color1 = color1 , color2 = color2, spaceMethod = spaceMethod)
 end
+
+function instant(green::Union{Green2DLR{DT,TT,TGT,SGT},GreenSym2DLR{DT,TT,TGT,SGT}}, space, spaceMethod) where {DT,TT,TGT,SGT}
+    return instant(; green=green, space = space, color1 = 1 , color2 = 1, spaceMethod = spaceMethod)
+end
+
+
 """
-    function dynamic(green::Green2DLR{DT,TT,TGT,SGT}, time, space, color1::Int, color2::Int, timeMethod::TM , spaceMethod::SM) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,TM,SM}
+    function dynamic(green::Union{Green2DLR{DT,TT,TGT,SGT},GreenSym2DLR{DT,TT,TGT,SGT}}, time, space, color1::Int, color2::Int, timeMethod::TM , spaceMethod::SM) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,TM,SM}
 
 Find value of Green's function's dynamic part at given color and k/x by interpolation.
 Interpolation method is by default depending on the grid, but could also be chosen to be linear.
@@ -365,16 +371,14 @@ Interpolation method is by default depending on the grid, but could also be chos
 - 'timeMethod': Method of interpolation for time
 - 'spaceMethod': Method of interpolation for space 
 """
-function dynamic(green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, color1::Int, color2::Int, timeMethod::TM , spaceMethod::SM) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,TM,SM}
+function _dynamic(TIM, SIM; green::Union{Green2DLR{DT,TT,TGT,SGT},GreenSym2DLR{DT,TT,TGT,SGT}}, time, space, color1::Int, color2::Int) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid}
     # for double composite
     if isempty(green.dynamic)
         error("Dynamic Green's function can not be empty!")
     else
-        TIM = InterpMethod(TGT,SM)
-        SIM = InterpMethod(SGT,SM)
-
         spaceNeighbor = CompositeGrids.Interp.findneighbor(SIM, green.spaceGrid, space)
-        if green.timeType == ImFreq && TM != DLRINTERP
+        println(TIM)
+        if green.timeType == ImFreq && TIM != DLRInterp
             timeGrid = (green.timeGrid.grid * 2 .+ 1) * π / green.β
             comTimeGrid = CompositeGrids.SimpleG.Arbitrary{eltype(timeGrid)}(timeGrid)
             comTime = (2*time+1)*π/green.β
@@ -385,20 +389,14 @@ function dynamic(green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, color1::Int, c
 
         timeNeighbor = CompositeGrids.Interp.findneighbor(TIM, comTimeGrid, comTime)
         dynamic_slice = view(green.dynamic, color1, color2, spaceNeighbor.index, timeNeighbor.index)
-
         dynamic_slice_xint = CompositeGrids.Interp.interpsliced(spaceNeighbor,dynamic_slice, axis=1)
         result = CompositeGrids.Interp.interpsliced(timeNeighbor,dynamic_slice_xint, axis=1)
     end
     return result
 end
 
-function dynamic(green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space,  timeMethod::TM , spaceMethod::SM) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,TM,SM}
-    return  dynamic(green, time, space, 1, 1,  timeMethod , spaceMethod)
-end
-
-function dynamic(
-    green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, color1::Int, color2::Int,
-    timeMethod::LinearInterp = LINEARINTERP , spaceMethod::LinearInterp = LINEARINTERP 
+function _dynamic( ::LinearInterp , ::LinearInterp
+    ;green::Union{Green2DLR{DT,TT,TGT,SGT},GreenSym2DLR{DT,TT,TGT,SGT}}, time, space, color1::Int, color2::Int,
     ) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid}
     # for double composite and double linear
     if isempty(green.dynamic)
@@ -413,29 +411,19 @@ function dynamic(
             comTime = time
         end
         dynamic_slice = view(green.dynamic, color1, color2, :,:)
-        result = CompositeGrids.Interp.linear2D(dynamic_slice, green.spaceGrid,comTimeGrid,space,comTime)
+        result = CompositeGrids.Interp.linear2D(dynamic_slice, green.spaceGrid, comTimeGrid,space,comTime)
     end
     return result
 end
 
-function dynamic(
-    green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space,
-    timeMethod::LinearInterp = LINEARINTERP , spaceMethod::LinearInterp = LINEARINTERP
-    ) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid}
-    return dynamic( green, time, space, 1, 1, timeMethod, spaceMethod)
-end
 
-
-
-function dynamic(green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, color1::Int, color2::Int, timeMethod::DLRInterp , spaceMethod::SM) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,SM}
+function _dynamic(::DLRInterp, SIM; green::Union{Green2DLR{DT,TT,TGT,SGT},GreenSym2DLR{DT,TT,TGT,SGT}}, time, space, color1::Int, color2::Int) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid}
     # for composite space and dlr time
     if isempty(green.dynamic)
         error("Dynamic Green's function can not be empty!")
     else
-        SIM = InterpMethod(SGT,SM)
         spaceNeighbor = CompositeGrids.Interp.findneighbor(SIM, green.spaceGrid, space)
         dynamic_slice = view(green.dynamic, color1, color2, spaceNeighbor.index,:)
-
         dynamic_slice_xint = CompositeGrids.Interp.interpsliced(spaceNeighbor,dynamic_slice, axis=1)
         if green.timeType == ImFreq
             result = (matfreq2matfreq(green.dlrGrid, dynamic_slice_xint, [time,], green.timeGrid.grid))[1]
@@ -446,6 +434,85 @@ function dynamic(green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, color1::Int, c
     return result
 end
 
-function dynamic(green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, timeMethod::DLRInterp , spaceMethod::SM) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,SM}
-    return dynamic(green, time, space, 1, 1, timeMethod, spaceMethod)
+dynamic(;timeMethod::TM, spaceMethod::SM ,green::Union{Green2DLR{DT,TT,TGT,SGT},GreenSym2DLR{DT,TT,TGT,SGT}}, time, space , color1::Int, color2::Int) where {TM,SM,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,DT,TT} = _dynamic(InterpMethod(TGT,TM), InterpMethod(SGT, SM); green, time, space, color1, color2)
+
+function dynamic(green::Union{Green2DLR{DT,TT,TGT,SGT},GreenSym2DLR{DT,TT,TGT,SGT}}, time, space, color1::Int=1, color2::Int=color1, timeMethod::TM=DEFAULTINTERP , spaceMethod::SM=DEFAULTINTERP) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,TM,SM}
+    return  dynamic(; timeMethod = timeMethod, spaceMethod = spaceMethod, green = green, time=time, space=space, color1=color1, color2 =color2)
 end
+
+function dynamic(green::Union{Green2DLR{DT,TT,TGT,SGT},GreenSym2DLR{DT,TT,TGT,SGT}}, time, space, timeMethod::TM, spaceMethod::SM) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,TM,SM}
+    return  dynamic(; timeMethod = timeMethod, spaceMethod = spaceMethod, green = green, time=time, space=space, color1=1, color2 =1)
+end
+
+# function _dynamic(TIM, SIM; green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, color1::Int, color2::Int) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid}
+#     # for double composite
+#     if isempty(green.dynamic)
+#         error("Dynamic Green's function can not be empty!")
+#     else
+#         spaceNeighbor = CompositeGrids.Interp.findneighbor(SIM, green.spaceGrid, space)
+#         println(TIM)
+#         if green.timeType == ImFreq && TIM != DLRInterp
+#             timeGrid = (green.timeGrid.grid * 2 .+ 1) * π / green.β
+#             comTimeGrid = CompositeGrids.SimpleG.Arbitrary{eltype(timeGrid)}(timeGrid)
+#             comTime = (2*time+1)*π/green.β
+#         else
+#             comTimeGrid = green.timeGrid
+#             comTime = time
+#         end
+
+#         timeNeighbor = CompositeGrids.Interp.findneighbor(TIM, comTimeGrid, comTime)
+#         dynamic_slice = view(green.dynamic, color1, color2, spaceNeighbor.index, timeNeighbor.index)
+#         dynamic_slice_xint = CompositeGrids.Interp.interpsliced(spaceNeighbor,dynamic_slice, axis=1)
+#         result = CompositeGrids.Interp.interpsliced(timeNeighbor,dynamic_slice_xint, axis=1)
+#     end
+#     return result
+# end
+
+# function _dynamic( ::LinearInterp , ::LinearInterp
+#     ;green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, color1::Int, color2::Int,
+#     ) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid}
+#     # for double composite and double linear
+#     if isempty(green.dynamic)
+#         error("Dynamic Green's function can not be empty!")
+#     else
+#         if green.timeType == ImFreq
+#             timeGrid = (green.timeGrid.grid * 2 .+ 1) * π / green.β
+#             comTimeGrid = CompositeGrids.SimpleG.Arbitrary{eltype(timeGrid)}(timeGrid)            
+#             comTime = (2*time+1)*π/green.β
+#         else
+#             comTimeGrid = green.timeGrid
+#             comTime = time
+#         end
+#         dynamic_slice = view(green.dynamic, color1, color2, :,:)
+#         result = CompositeGrids.Interp.linear2D(dynamic_slice, green.spaceGrid, comTimeGrid,space,comTime)
+#     end
+#     return result
+# end
+
+
+# function _dynamic(::DLRInterp, SIM; green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, color1::Int, color2::Int) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid}
+#     # for composite space and dlr time
+#     if isempty(green.dynamic)
+#         error("Dynamic Green's function can not be empty!")
+#     else
+#         spaceNeighbor = CompositeGrids.Interp.findneighbor(SIM, green.spaceGrid, space)
+#         dynamic_slice = view(green.dynamic, color1, color2, spaceNeighbor.index,:)
+#         dynamic_slice_xint = CompositeGrids.Interp.interpsliced(spaceNeighbor,dynamic_slice, axis=1)
+#         if green.timeType == ImFreq
+#             result = (matfreq2matfreq(green.dlrGrid, dynamic_slice_xint, [time,], green.timeGrid.grid))[1]
+#         elseif green.timeType == ImTime
+#             result = (tau2tau(green.dlrGrid, dynamic_slice_xint, [time,], green.timeGrid.grid))[1]
+#         end
+#     end
+#     return result
+# end
+
+# dynamic(;timeMethod::TM, spaceMethod::SM ,green::Union{GreenSym2DLR{DT,TT,TGT,SGT}}, time, space , color1::Int, color2::Int) where {TM,SM,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,DT,TT} = _dynamic(InterpMethod(TGT,TM), InterpMethod(SGT, SM); green, time, space, color1, color2)
+
+# function dynamic(green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, color1::Int=1, color2::Int=color1, timeMethod::TM=DEFAULTINTERP , spaceMethod::SM=DEFAULTINTERP) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,TM,SM}
+#     return  dynamic(; timeMethod = timeMethod, spaceMethod = spaceMethod, green = green, time=time, space=space, color1=color1, color2 =color2)
+# end
+
+# function dynamic(green::GreenSym2DLR{DT,TT,TGT,SGT}, time, space, timeMethod::TM, spaceMethod::SM) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,TM,SM}
+#     return  dynamic(; timeMethod = timeMethod, spaceMethod = spaceMethod, green = green, time=time, space=space, color1=1, color2 =1)
+# end
