+++ test/runtests.jl
@@ -4,7 +4,7 @@ using JLD2, FileIO
 using CodecZlib
 
 if isempty(ARGS)
-    include("test_Green.jl")
+    #include("test_Green.jl")
     include("test_Green_sym.jl")
     #include("interpolate.jl")
 else
